# server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_AboloreDev_geritcht-restaurant_internals_dto.CreateReservationRequest:
    properties:
      date:
        type: string
      party_size:
        type: integer
      special_requests:
        type: string
      table_id:
        type: integer
      time_slot:
        type: string
    required:
    - date
    - party_size
    - table_id
    - time_slot
    type: object
...
  github_com_AboloreDev_geritcht-restaurant_internals_dto.ReservationResponse:
    properties:
      checked_in_at:
        type: string
      created_at:
        type: string
      date:
        type: string
      id:
        type: integer
      party_size:
        type: integer
      special_requests:
        type: string
      status:
        type: string
      table:
        $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_dto.TableResponse'
      table_id:
        type: integer
      time_slot:
        type: string
      user:
        $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_dto.UserResponse'
      user_id:
        type: integer
    type: object
...
  github_com_AboloreDev_geritcht-restaurant_internals_dto.TableResponse:
    properties:
      capacity:
        type: integer
      id:
        type: integer
      location:
        type: string
      name:
        type: string
      qr_code:
        type: string
      status:
        type: string
    type: object
...
  github_com_AboloreDev_geritcht-restaurant_internals_dto.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      email_verified:
        type: boolean
      first_name:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      last_name:
        type: string
      phone_number:
        type: string
      role:
        type: string
    type: object
...
  github_com_AboloreDev_geritcht-restaurant_internals_utils.PaginatedMeta:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  github_com_AboloreDev_geritcht-restaurant_internals_utils.PaginatedResponse:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      meta:
        $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.PaginatedMeta'
      status:
        type: boolean
    type: object
  github_com_AboloreDev_geritcht-restaurant_internals_utils.Response:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      status:
        type: boolean
    type: object
host: https://geritcht-restaurant.onrender.com
info:
  contact:
    email: alabifathiu@yahoo.com
    name: Alabi Fathiu
    url: http://linkedin.com/in/fathiu-alabi
  description: A modern Restaurant API built with Go, Gin, and GORM
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Geritcht Restaurant API
  version: "1.0"
paths:
...
  /availability:
    get:
      consumes:
      - application/json
      description: Check available tables for a specified date, time, and party size.
      parameters:
      - description: Reservation date (YYYY-MM-DD)
        in: query
        name: date
        required: true
        type: string
      - description: Reservation time (HH:MM)
        in: query
        name: time
        required: true
        type: string
      - description: Number of guests
        in: query
        name: guests
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Availability retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_dto.TableResponse'
                  type: array
              type: object
        "400":
          description: Invalid request data or time slot
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
      summary: Check table availability
      tags:
      - Reservations
...
  /reservations:
    get:
      consumes:
      - application/json
      description: Retrieve all reservations. Admin access required.
      parameters:
      - description: Reservation status
        in: query
        name: status
        type: string
      - description: Start date (YYYY-MM-DD)
        in: query
        name: from
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Reservations retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_dto.ReservationResponse'
                  type: array
              type: object
        "400":
          description: Invalid request data
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
      security:
      - BearerAuth: []
      summary: List all reservations
      tags:
      - Reservations
    post:
      consumes:
      - application/json
      description: Create a new table reservation for the authenticated user.
      parameters:
      - description: Reservation details
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_dto.CreateReservationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Reservation created successfully
          headers:
            X-RateLimit-Limit:
              description: Maximum requests allowed
              type: string
            X-RateLimit-Remaining:
              description: Remaining requests in the current window
              type: string
          schema:
            allOf:
            - $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_dto.ReservationResponse'
              type: object
        "400":
          description: Invalid request data, invalid time slot, table unavailable,
            or invalid table capacity
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "404":
          description: Table not found
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "429":
          description: Too many requests
          headers:
            X-RateLimit-Limit:
              description: Maximum requests allowed
              type: string
            X-RateLimit-Remaining:
              description: Remaining requests (0 when limited)
              type: string
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
      security:
      - BearerAuth: []
      summary: Create a reservation
      tags:
      - Reservations
  /reservations/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve a reservation by its ID for the authenticated user.
      parameters:
      - description: Reservation ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Reservation retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_dto.ReservationResponse'
              type: object
        "400":
          description: Invalid reservation ID
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "404":
          description: Reservation not found
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
      security:
      - BearerAuth: []
      summary: Get reservation
      tags:
      - Reservations
  /reservations/{id}/cancel:
    patch:
      consumes:
      - application/json
      description: Cancel an existing reservation owned by the authenticated user.
      parameters:
      - description: Reservation ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Reservation cancelled successfully
          headers:
            X-RateLimit-Limit:
              description: Maximum requests allowed
              type: string
            X-RateLimit-Remaining:
              description: Remaining requests in the current window
              type: string
          schema:
            allOf:
            - $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_dto.ReservationResponse'
              type: object
        "400":
          description: Reservation cannot be cancelled
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "404":
          description: Reservation or user not found
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "429":
          description: Too many requests
          headers:
            X-RateLimit-Limit:
              description: Maximum requests allowed
              type: string
            X-RateLimit-Remaining:
              description: Remaining requests (0 when limited)
              type: string
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
      security:
      - BearerAuth: []
      summary: Cancel reservation
      tags:
      - Reservations
  /reservations/{id}/check-in:
    post:
      consumes:
      - application/json
      description: Check in the authenticated user for an existing reservation.
      parameters:
      - description: Reservation ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Checked in successfully
          headers:
            X-RateLimit-Limit:
              description: Maximum requests allowed
              type: string
            X-RateLimit-Remaining:
              description: Remaining requests in the current window
              type: string
          schema:
            allOf:
            - $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_dto.ReservationResponse'
              type: object
        "400":
          description: Cannot check in or reservation already checked in
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "404":
          description: Reservation not found
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "429":
          description: Too many requests
          headers:
            X-RateLimit-Limit:
              description: Maximum requests allowed
              type: string
            X-RateLimit-Remaining:
              description: Remaining requests (0 when limited)
              type: string
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
      security:
      - BearerAuth: []
      summary: Check in reservation
      tags:
      - Reservations
  /reservations/search:
    get:
      consumes:
      - application/json
      description: Search reservations by name with pagination
      parameters:
      - description: Reservation name
        in: query
        name: query
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Number of items per page
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Reservations retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_dto.ReservationResponse'
                  type: array
              type: object
        "400":
          description: Invalid search parameters
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "404":
          description: No reservations found
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
      security:
      - BearerAuth: []
      summary: Search Reservations
      tags:
      - Reservations
  /reservations/today:
    get:
      consumes:
      - application/json
      description: Retrieve all reservations for today. Admin access required.
      parameters:
      - description: Reservation status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Reservations retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_dto.ReservationResponse'
                  type: array
              type: object
        "400":
          description: Invalid request data
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
      security:
      - BearerAuth: []
      summary: Get today's reservations
      tags:
      - Reservations
  /reservations/user:
    get:
      consumes:
      - application/json
      description: Retrieve all reservations belonging to the authenticated user.
      parameters:
      - description: Reservation status
        in: query
        name: status
        type: string
      - description: Start date (YYYY-MM-DD)
        in: query
        name: from
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Reservations retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_dto.ReservationResponse'
                  type: array
              type: object
        "400":
          description: Invalid request data
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
      security:
      - BearerAuth: []
      summary: Get my reservations
      tags:
      - Reservations
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"